refactor(#17): ajustes
