+++ src/middlewares/AuthMiddleware.js
@@ -11,8 +11,7 @@ module.exports = (req, res, next) => {
 
   const parts = authHeader.split(" ");
 
-  if (!parts.length === 2)
-    return res.status(401).json({ error: "Token error" });
+  if (parts.length !== 2) return res.status(401).json({ error: "Token error" });
 
   const [schema, token] = parts;
